Выносит функции для работы с DOM-элементами в отдельный файл
